Implement page creation / deletion

+++ app/service/page.go
@@ -34,6 +34,10 @@ func (s *Page) GetAll() ([]database.Page, error) {
 }
 
 func (s *Page) Delete(id string) error {
+	if id == "index" {
+		return fmt.Errorf("can't delete index page")
+	}
+
 	if err := s.db.DeletePage(id); err != nil {
 		return fmt.Errorf("failed to delete page by id %s: %v", id, err)
 	}
